# Decide inline-worthiness based on a more nuanced cost model