@timholy 2017-06-04
Just realized there's a really big hole in this: a `for` or `while` loop will only get counted as if its body is executed once. Probably it would be best to just not inline functions with such loops. I don't see a "loop detector" presently, but perhaps I'm just missing it.