@timholy 2017-06-06
OK, I've taken this further; a modified version of this (that had some debugging code) locally passed all tests but one. For demonstration purposes, I've also pushed a branch, `teh/disable_inline_macros`, that gets rid of all forced-inlining in Base; the point of this is to force the inliner to "stand on its own" and see what happens. I then rebased this PR on that branch. My original intention was to then run nanosoldier to compare the differences, but from my local testing there are so many differences that it's a bit overwhelming.

The bottom line seems pretty simple, though. Behold:
![inline_perf](https://user-images.githubusercontent.com/1525481/26852137-e6730dec-4ad2-11e7-8b30-a4f0100f8657.png)

So while there are a few regressions, they are both outnumbered and outweighed by the improvements.

As one example, `map(abs, t)` where `t = ntuple(identity, n)` inlines (on its own) up to `n=3` on master but `n=14` on this branch; a more expensive function, `map(sqrt, t)`, inlines up to `n=5`.

Now, one concern might be that this simply inlines everything, and that's cheating. What's the best way to assess such concerns? I checked the "resident size" in `top`, and compared to 0.6 (I didn't feel like rebuilding master) the size was pretty comparable.