@yuyichao 2017-06-06
Am I reading it correctly that there's one case that shows 10^6 speed up?

> Now, one concern might be that this simply inlines everything, and that's cheating.

Well, then all optimizing compilers have been cheating all alone.... I'm pretty sure most C/C++ compiler do much more aggressive inlining than we do so I wouldn't call it cheating. The important thing is to measure the overhead. Maybe measuring the time to compile sysimg, or some other standard precompilation test cases. You can also measure the total size in inferenced IR record a statistic of non-void expression in inference after finishing each functions and if you want more low level info you can also check `size -A` of sysimg and record the JIT code size by adding a counter in `RTDyldMemoryManagerJL::allocateCodeSection` in `cgmemmgr.cpp`.