@JeffBezanson 2017-06-06
Wow, that's really cool. If you re-enable manual inlining declarations, do the regressions go away?