@timholy 2017-06-06
> Am I reading it correctly that there's one case that shows 10^6 speed up?

Yes, that's basically right; several others cases get better than 10^4 speed up. Or rather, "not slowed down"; I am sure that our performance annotations take care of all the egregious cases. I very much doubt we'll see much improvement compared to "real" master from this, but of course it's worth testing.

> Maybe measuring the time to compile sysimg

Great idea.

| task | master | PR |
|:-------:|:----------:|:------:|
| build from scratch (incl. `inference.jl`) | 2min 46s | 3min 6s |
| build `base/precompile.jl` | 1min 36s | 1min 50s |

So there's definitely a slowdown with this. Would be interested in people's thoughts on how bad this is.

> If you re-enable manual inlining declarations, do the regressions go away?

Will test. But right now I'm testing how this fares against "real" master, with all of its performance annotations; I'm curious to know if we could scale back on our usage of `@inline`.