@yuyichao 2017-06-06
I believe we also have a `ENABLE_TIMING` option in `options.h` which can give you a breakdown of the timing. That timing different doesn't sound too bad to me though compile time is also not the thing I care about the most....