@timholy 2017-06-07
Here's a comparison against master with all of its `@inline` annotations:

![inline_vs_master](https://user-images.githubusercontent.com/1525481/26860267-b39fbebc-4b02-11e7-9566-c03d8319beb2.png)

Definitely worse on balance, but better than I expected: to my eye, this branch *without* performance annotations is closer to master *with* annotations than it is to master with the annotations stripped out.

Obviously some of the regressions are pretty severe, but I've done literally no performance tuning here. Probably worth looking into some of the regressions by hand and seeing what I can come up with...in particular, the ones that have ~10-100ns of runtime on master seem likely to be just me assigning the wrong cost and it therefore failing to inline a pretty simple function.