@timholy 2017-06-08
@quinnj, the cost model is almost certainly not accurate enough to expose to users; it doesn't take into account looping, and the cost of non-inferrability is handled crudely. 

However, it can be useful to understand why a particular function didn't get inlined. I'll document how to call `statement_cost` for a function body.