@timholy 2017-06-09
So, I'm thinking of putting this on the back burner until #21888 merges; I'm seeing quite a few issues that I hope get resolved by better GC marking. The last commit here is just one example; I'm also seeing a few big performance regressions that seem to occur because we allocate when we don't need to; a good example is `mean(rand(10^3))`, where I'm getting allocations due to unnecessarily boxing the [`blksize`](https://github.com/JuliaLang/julia/blob/a52a78d3f8a1e12b9586dab8c5d1b3ed7c626126/base/reduce.jl#L177) argument. I'm pretty sure this is not caused by my changes here, but it does get triggered by these changes.

Of course the question is whether its benefits on LLVM 3.9 will fix these issues, or whether it will have to wait until we move to LLVM 5.