@timholy 2017-06-09
Speaking of bugs triggered by this, https://travis-ci.org/JuliaLang/julia/builds/240808331 hit an assertion in `gf.c`. (It was the `Int` version that was a subtype of the `Int...` version of that function.) I worked around it by declaring that argument as `Vararg{Int,N}`, but it suggests a bug elsewhere.