@timholy 2017-06-18
@nanosoldier `runbenchmarks(ALL, vs = ":master")`