@nanosoldier 2017-06-18
[Your benchmark job](https://github.com/JuliaLang/julia/pull/22210#issuecomment-309283556) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/77af818db5d74c448f55c75b4368653181a80b0d/37a9727_vs_e9f2072/report.md). cc @jrevels