@timholy 2017-06-18
That is a pretty darn satisfying number of green checkboxes on that nanosoldier run. In graphical format, here are my local results:
![inlining](https://user-images.githubusercontent.com/1525481/27263533-bdca1fea-5430-11e7-93ff-5b352709c511.png)

I still have a couple of errors to fix, though. This has the habit of triggering assertions in the C code when building with `FORCE_ASSERTIONS=1` (which is how Travis builds it). The latest is 
```
julia: /home/tim/src/julia-master/src/typemap.c:843: jl_typemap_level_assoc_exact: Assertion `(((jl_value_t*)(((jl_taggedvalue_t*)((char*)(ty) - sizeof(jl_taggedvalue_t)))->header & ~(uintptr_t)15))==(jl_value_t*)(jl_datatype_type))' failed.

signal (6): Aborted
while loading /home/tim/src/julia-master/test/linalg/diagonal.jl, in expression starting on line 10
```
If anyone has any ideas, I'm all ears.