@vtjnash 2017-06-26
This change also makes the sysimg about 50% larger and generates about 50% more code (see `size -A` results in the CI logs). That would generally seems to indicate this is just being much too aggressive at inlining, and not necessarily much more intelligent.